convert to float in math package

+++ src/main/java/math/Vector2f.kt
@@ -6,14 +6,14 @@ import kotlin.math.sin
 import kotlin.math.sqrt
 
 class Vector2f(
-    private var x: Double,
-    private var y: Double
+    private var x: Float,
+    private var y: Float
 ) : MatrixBuilder(arrayOf(arrayOf(x, y))) {
 
     val length = 2
 
     // auxiliary constructor
-    constructor(value: Double) : this(value, value)
+    constructor(value: Float) : this(value, value)
 
     fun normalize() {
         val length = sqrt(x.pow(2) + y.pow(2))
@@ -21,25 +21,25 @@ class Vector2f(
         y /= length
     }
 
-    fun rotate(theta: Double) {
+    fun rotate(theta: Float) {
         x = x * cos(theta) - y * sin(theta)
         y = x * sin(theta) + y * cos(theta)
     }
 
     // abstract members' implementation
-    override fun plusAssign(scalar: Double) {
+    override fun plusAssign(scalar: Float) {
         this += Vector2f(scalar)
     }
 
-    override fun minusAssign(scalar: Double) {
+    override fun minusAssign(scalar: Float) {
         this -= Vector2f(scalar)
     }
 
-    override fun plus(scalar: Double) = (this + Vector2f(scalar))!!
+    override fun plus(scalar: Float) = (this + Vector2f(scalar))!!
 
-    override fun minus(scalar: Double) = (this - Vector2f(scalar))!!
+    override fun minus(scalar: Float) = (this - Vector2f(scalar))!!
 
-    override fun inc() = this + 1.0
+    override fun inc() = this + 1.0f
 
-    override fun dec() = this - 1.0
+    override fun dec() = this - 1.0f
 }
